move dropdown executionContext and udapp related logic to blockchain model
